+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### :wrench: Changed
 
-- Display all cURL commands for stateful failures
+- Display all cURL commands for stateful failures.
+- Inject unconstrained path parameters when missing in the operation definition (schema error still reported).
 
 ## [4.1.0](https://github.com/schemathesis/schemathesis/compare/v4.0.26...v4.1.0) - 2025-08-17